% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oabp_mission}
\alias{print.oabp_mission}
\title{Print an OABP mission}
\usage{
\method{print}{oabp_mission}(x, ...)
}
\arguments{
\item{x}{An \code{oabp_mission} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an OABP mission
}
