% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oabp_stats}
\alias{print.oabp_stats}
\title{Print OABP protocol statistics}
\usage{
\method{print}{oabp_stats}(x, ...)
}
\arguments{
\item{x}{An \code{oabp_stats} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print OABP protocol statistics
}
